+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+* Add an `--ignore-errors` CLI argument, config value, & env var that logs API
+  errors & continues processing instead of halting execution.
+
 
 ## v0.1.0.2